requirements.txt
bs4
glfw
imgui
numpy
opencv-python
psutil
pyinjector
pillow
PyGitHub
PyOpenGL
PyOpenGL_accelerate
pywin32
requests
requests_cache
win11toast
